requirements.txt
paddleocr>=2.6.0
paddlepaddle>=2.5.0
gradio>=3.50.2
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
openpyxl>=3.0.0